Fix: SSE 통신 과정에서 발생 가능한 예외 처리 추가 #78
- timeout을 지정하여, 10분 동안 메시지 요청이 없을 경우 연결을 끊고 클라이언트 측에서 재연결을 요청하는 기능 추가
- 메시지 전송 대상을 찾지 못했을 경우 404 Not Found
- 전송 대상을 찾았으나, 전송에 실패했을 경우 500 Server Error
